SOLUTION_real_ai_images.py
"""
SOLUTION: Use VS Code's MCP tool directly through the extension API

Since the MCP tool works in VS Code but not via subprocess, we need to:
1. Use the image generation that IS available: OpenAI DALL-E or similar
2. OR: Create a VS Code extension endpoint that wraps the MCP call
3. OR: Use the Gemini API correctly with available models

Let's test option 3: Using Gemini API with correct endpoints
"""

import os
import base64
import requests
from pathlib import Path
from dotenv import load_dotenv
from PIL import Image
import io

load_dotenv()

# SOLUTION: The MCP tool in VS Code uses Gemini's INTERNAL image generation
# which is NOT exposed via the public API yet. 
# We have 3 real options:

print("="*70)
print("ANALYSIS: Why MCP works in VS Code but not in Python")
print("="*70)

print("""
❌ PROBLEM: The @modelcontextprotocol/server-gemini-2-5-flash-image 
   is NOT an npm package - it's built into VS Code's Copilot extension.

✅ SOLUTIONS:

1. Use OpenAI DALL-E API (works, paid)
2. Use Stable Diffusion API (free options available)
3. Use Hugging Face image models (free)
4. Wait for Google to release public Imagen API access

RECOMMENDATION: Use Hugging Face's free Stable Diffusion endpoint
""")

def test_huggingface_image():
    """Test Hugging Face image generation - FREE and works!"""
    
    API_URL = "https://api-inference.huggingface.co/models/stabilityai/stable-diffusion-xl-base-1.0"
    
    # You need a HuggingFace token (free): https://huggingface.co/settings/tokens
    HF_TOKEN = os.getenv('HUGGINGFACE_TOKEN', '')
    
    if not HF_TOKEN:
        print("\n⚠️  Need HuggingFace token. Get one FREE at:")
        print("   https://huggingface.co/settings/tokens")
        print("\n   Then add to .env: HUGGINGFACE_TOKEN=your_token")
        return False
    
    headers = {"Authorization": f"Bearer {HF_TOKEN}"}
    
    prompt = "professional modern office workspace, laptop on desk, natural lighting, clean aesthetic, photorealistic"
    
    print(f"\n🎨 Generating image with Stable Diffusion XL...")
    print(f"📝 Prompt: {prompt}")
    
    try:
        response = requests.post(
            API_URL,
            headers=headers,
            json={"inputs": prompt},
            timeout=60
        )
        
        if response.status_code == 200:
            # Save image
            output = Path(__file__).parent / "test_output" / "hf_test.png"
            output.parent.mkdir(exist_ok=True)
            
            image = Image.open(io.BytesIO(response.content))
            image.save(output)
            
            print(f"✅ Image generated: {output.stat().st_size:,} bytes")
            print(f"📁 Saved to: {output}")
            return True
        else:
            print(f"❌ Error {response.status_code}: {response.text}")
            return False
            
    except Exception as e:
        print(f"❌ Error: {e}")
        return False

if __name__ == "__main__":
    print("\n" + "="*70)
    print("TESTING: HuggingFace Stable Diffusion (FREE)")
    print("="*70)
    
    success = test_huggingface_image()
    
    print("\n" + "="*70)
    if success:
        print("✅ SUCCESS - We can use this for real AI images!")
    else:
        print("ℹ️  Setup needed - but this is the solution")
    print("="*70)
