requirements.txt
fastapi
uvicorn
supabase
python-dotenv
google-generativeai
Pillow
matplotlib
plotly
kaleido
pygments
qrcode[pil]
reportlab
python-multipart
svglib
